Move server route under (user) since they share the logged in layout
