+++ app/(user)/server/[serverId]/layout.tsx
@@ -1,10 +1,7 @@
-
 import { Params } from "next/dist/shared/lib/router/utils/route-matcher";
-import ColumnWrapper from "../../_components/wrappers/ColumnWrapper";
-import TopMenu from "@/app/_components/TopMenu";
+import LayoutClientWrapper from "./_components/LayoutClientWrapper";
 import ServerInnerNav from "./_components/ServerInnerNav";
 import ServerMembersMenu from "./_components/ServerMembersMenu";
-import LayoutClientWrapper from "./_components/LayoutClientWrapper";
 
 export default function ServerLayout({
   params,
@@ -16,10 +13,12 @@ export default function ServerLayout({
   const id = params.serverId;
 
   return (
-    <LayoutClientWrapper id={id} innerNav={<ServerInnerNav id={id} />} membersMenu={<ServerMembersMenu id={id}/>}>
+    <LayoutClientWrapper
+      id={id}
+      innerNav={<ServerInnerNav id={id} />}
+      membersMenu={<ServerMembersMenu id={id} />}
+    >
       {children}
     </LayoutClientWrapper>
-    
-    
   );
 }
