+++ app/(user)/server/[serverId]/loading.tsx
@@ -1,0 +1,9 @@
+import ColumnWrapper from "@/app/_components/wrappers/ColumnWrapper";
+
+export default function ServerHomeSuspense() {
+  return (
+    <div className="flex flex-grow">
+      <ColumnWrapper className="sticky mr-2 h-full w-[15%] border-r border-r-black50 p-0"></ColumnWrapper>
+    </div>
+  );
+}
