+++ app/(user)/server/[serverId]/page.tsx
@@ -1,7 +1,7 @@
+import FeedbackCard from "@/app/_components/FeedbackCard";
+import Main from "@/app/_components/wrappers/PageMain";
+import { getServerData } from "@/prisma/services/serverService";
 import { Params } from "next/dist/shared/lib/router/utils/route-matcher";
-import Main from "../../_components/wrappers/PageMain";
-import { getServerData } from "../../../prisma/services/serverService";
-import FeedbackCard from "../../_components/FeedbackCard";
 
 export default async function ServerHome({ params }: { params: Params }) {
   const id = params.serverId;
