+++ app/(user)/server/join/[invitationId]/_components/NotProtected.tsx
