+++ app/(user)/server/join/[invitationId]/_components/PasswordProtected.tsx
