+++ app/(user)/server/join/[invitationId]/page.tsx
